Modifying build.proj to not error on xunit execution returning not 0. It returns not zero when tests are skipped as well. Instead, let's rely on Jenkins validating the xml file.
Fixing the test that I implicit set to fail to test Jenkins xml test parsing. Also changing the name of the test projects to avoid conflict when using the files.
