+++ test/Microsoft.NET.Build.Tests/GivenThatWeWantToVerifyNuGetReferenceCompat.cs
@@ -16,6 +16,8 @@ namespace Microsoft.NET.Build.Tests
 {
     public class GivenThatWeWantToVerifyNuGetReferenceCompat : SdkTest, IClassFixture<DeleteNuGetArtifactsFixture>
     {
+        private TestPackageReference _net461PackageReference;
+
         [Theory]
         [InlineData("net45", "Full", "netstandard1.0 netstandard1.1 net45", true, true)]
         [InlineData("net451", "Full", "netstandard1.0 netstandard1.1 netstandard1.2 net45 net451", true, true)]
@@ -137,6 +139,115 @@ namespace Microsoft.NET.Build.Tests
             {
                 referencerBuildResult.Should().Fail().And.HaveStdOutContaining("It cannot be referenced by a project that targets");
             }
+        }
+
+        [WindowsOnlyTheoryAttribute]
+        [InlineData("netstandard2.0")]
+        [InlineData("netcoreapp2.0")]
+        public void Net461_is_implicit_for_Netstandard_and_Netcore_20(string targetFramework)
+        {
+            var testProjectName = targetFramework.Replace(".", "_") + "implicit_ptf";
+
+            var testProjectTestAsset = CreateTestAsset(testProjectName, targetFramework);
+
+            var restoreCommand = testProjectTestAsset.GetRestoreCommand(relativePath: testProjectName);
+            restoreCommand.AddSource(Path.GetDirectoryName(_net461PackageReference.NupkgPath));
+            restoreCommand.Execute().Should().Pass();
+
+            var buildCommand = new BuildCommand(
+                Stage0MSBuild,
+                Path.Combine(testProjectTestAsset.TestRoot, testProjectName));
+            buildCommand.Execute().Should().Pass();
+        }
+
+        [WindowsOnlyTheoryAttribute]
+        [InlineData("netstandard1.6")]
+        [InlineData("netcoreapp1.1")]
+        public void Net461_is_not_implicit_for_Netstandard_and_Netcore_less_than_20(string targetFramework)
+        {
+            var testProjectName = targetFramework.Replace(".", "_") + "non_implicit_ptf";
+
+            var testProjectTestAsset = CreateTestAsset(testProjectName, targetFramework);
+
+            var restoreCommand = testProjectTestAsset.GetRestoreCommand(relativePath: testProjectName);
+            restoreCommand.AddSource(Path.GetDirectoryName(_net461PackageReference.NupkgPath));
+            restoreCommand.Execute().Should().Fail();
+        }
+
+        [WindowsOnlyFact]
+        public void It_is_possible_to_disabled_net461_implicit_package_target_fallback()
+        {
+            const string testProjectName = "netstandard20_disabled_ptf";
+
+            var testProjectTestAsset = CreateTestAsset(
+                testProjectName,
+                "netstandard2.0",
+                new Dictionary<string, string> { {"DisableImplicitPackageTargetFallback", "true" } });
+
+            var restoreCommand = testProjectTestAsset.GetRestoreCommand(relativePath: testProjectName);
+            restoreCommand.AddSource(Path.GetDirectoryName(_net461PackageReference.NupkgPath));
+            restoreCommand.Execute().Should().Fail();
+        }
+
+        private TestAsset CreateTestAsset(
+            string testProjectName,
+            string targetFramework,
+            Dictionary<string, string> additionalProperties = null)
+        {
+            _net461PackageReference = CreateNet461Package(testProjectName);
+
+            var testProject =
+                new TestProject
+                {
+                    Name = testProjectName,
+                    TargetFrameworks = targetFramework,
+                    IsSdkProject = true
+                };
+
+            if (additionalProperties != null)
+            {
+                foreach (var additionalProperty in additionalProperties)
+                {
+                    testProject.AdditionalProperties.Add(additionalProperty.Key, additionalProperty.Value);    
+                }
+            }
+            
+            testProject.PackageReferences.Add(_net461PackageReference);
+
+            var testProjectTestAsset = _testAssetsManager.CreateTestProject(
+                testProject,
+                ConstantStringValues.TestDirectoriesNamePrefix, $"_{testProjectName}_net461");
+
+            return testProjectTestAsset;
+        }
+
+        private TestPackageReference CreateNet461Package(string testProjectName)
+        {
+            var net461Project = 
+                new TestProject
+                {
+                    Name = $"net461_{testProjectName}",
+                    TargetFrameworks = "net461",
+                    IsSdkProject = true
+                };
+
+            var net461PackageReference =
+                new TestPackageReference(
+                    net461Project.Name,
+                    "1.0.0",
+                    ConstantStringValues.ConstructNuGetPackageReferencePath(net461Project));
+
+            var net461PackageTestAsset = _testAssetsManager.CreateTestProject(
+                net461Project,
+                ConstantStringValues.TestDirectoriesNamePrefix,
+                ConstantStringValues.NuGetSharedDirectoryNamePostfix);
+            var packageRestoreCommand =
+                net461PackageTestAsset.GetRestoreCommand(relativePath: net461Project.Name).Execute().Should().Pass();
+            var dependencyProjectDirectory = Path.Combine(net461PackageTestAsset.TestRoot, net461Project.Name);
+            var packagePackCommand =
+                new PackCommand(Stage0MSBuild, dependencyProjectDirectory).Execute().Should().Pass();
+
+            return net461PackageReference;
         }
 
         TestProject GetTestProject(string name, string target, bool isSdkProject)
